Restructuring of files
