+++ COSC345-Project/Folder.swift
@@ -34,7 +34,7 @@ struct TagView: View {
 /// Extension of the Folder object
 extension Folder {
     /// Defines the folders utilized, with their respective icons and colours.
-    static let folderMatch = [(title: "Default", icon: "folder", color: "text"),
+    static let folderMatch = [(title: "Default", icon: "doc.plaintext", color: "text"),
                               (title: "Retail", icon: "tag", color: "blue"),
                               (title: "Groceries", icon: "cart", color: "green"),
                               (title: "Clothing", icon: "bag", color: "pink")]
