+++ COSC345-Project/Scanners.swift
@@ -10,11 +10,82 @@ import SwiftUI
 import VisionKit
 import Vision
 
-// GALLERY SCANNER
-struct GalleryScannerView: View {
+// CAMERA SCANNER
+struct DocumentScannerView: View {
+    @State var invalidAlert: Bool = false
     @Binding var scanSelection: ScanSelection
     @Binding var isRecognizing: Bool
-    @ObservedObject var recognizedContent: RecognizedContent = RecognizedContent()
+    @ObservedObject var recognizedContent: RecognizedContent  = RecognizedContent()
+    
+    var body: some View {
+        if !UIDevice.current.isSimulator { // if device is physical (supports camera)
+            DocumentScanner { result in
+                switch result {
+                    case .success(let scannedImages):
+                        isRecognizing = true
+                        print(recognizedContent.items.count)
+                        TextRecognition(scannedImages: scannedImages, recognizedContent: recognizedContent) {
+                            if saveReceipt(){ // if save receipt returns true (valid scan), exit the scanner
+                                scanSelection = .none
+                            } else { // else stay in the scanner and alert them to scan again
+                                invalidAlert = true
+                            }
+                            isRecognizing = false // Text recognition is finished, hide the progress indicator.
+                        }.recognizeText()
+                    case .failure(let error):
+                        print(error.localizedDescription)
+                }
+            } didCancelScanning: {
+                // Dismiss the scanner
+                scanSelection = .none
+            }.alert(isPresented: $invalidAlert) {
+                Alert(
+                    title: Text("Receipt Not Saved!"),
+                    message: Text("This image is not valid. Try again."),
+                    dismissButton: .default(Text("Okay"))
+                )
+            }
+        } else { // else if its in the simulator (no camera)
+            Text("Camera not supported in the simulator!\n\nPlease use a physical device.")
+                .font(.system(.title, design: .rounded))
+                .padding()
+            Button(action: {
+                scanSelection = .none
+            }){
+                Text("BACK")
+                    .padding().padding(.horizontal)
+                    .background(RoundedRectangle(cornerRadius: 25).fill(Color("object")))
+                    .padding()
+            }.buttonStyle(ShrinkingButton())
+            Spacer()
+        }
+    }
+    
+    /* i would put saveReceipt() in a separate function since its duplicated in both scanners,
+    but scanSelection keeps saying "im a let tho, you cant change me"
+    despite me assigning it as a Binding<ScanSelection> in the function params :( */
+    func saveReceipt() -> Bool{
+        if recognizedContent.items.count > 0 {
+            for receipt in recognizedContent.items {
+                if receipt.text.count < 2 {
+                    return false
+                } else {
+                    Receipt.saveScan(recognizedText: receipt.text)
+                    return true
+                }
+            }
+            scanSelection = .none
+        }
+        return false
+    }
+}
+
+// GALLERY SCANNER
+struct GalleryScannerView: View {
+    @State var invalidAlert: Bool = false
+    @Binding var scanSelection: ScanSelection
+    @Binding var isRecognizing: Bool
+    @ObservedObject var recognizedContent: RecognizedContent  = RecognizedContent()
     
     var body: some View {
         ImagePicker { result in
@@ -23,28 +94,44 @@ struct GalleryScannerView: View {
                     isRecognizing = true
                     print(recognizedContent.items.count)
                     TextRecognition(scannedImages: scannedImages, recognizedContent: recognizedContent) {
-                        saveReceipt()
+                        if saveReceipt(){ // if save receipt returns true (valid scan), exit the scanner
+                            scanSelection = .none
+                        } else { // else stay in the scanner and alert them to scan again
+                            invalidAlert = true
+                        }
                         isRecognizing = false // Text recognition is finished, hide the progress indicator.
-                        print(recognizedContent.items[0])
                     }.recognizeText()
                 case .failure(let error):
                     print(error.localizedDescription)
             }
-            scanSelection = .none
         } didCancelScanning: {
             // Dismiss the scanner
             scanSelection = .none
+        }.alert(isPresented: $invalidAlert) {
+            Alert(
+                title: Text("Receipt Not Saved!"),
+                message: Text("This image is not valid. Try again."),
+                dismissButton: .default(Text("Okay"))
+            )
         }
     }
     
-    func saveReceipt(){
+    /* i would put saveReceipt() in a separate function since its duplicated in both scanners,
+    but scanSelection keeps saying "im a let tho, you cant change me"
+    despite me assigning it as a Binding<ScanSelection> in the function params :( */
+    func saveReceipt() -> Bool{
         if recognizedContent.items.count > 0 {
             for receipt in recognizedContent.items {
-                Receipt.saveScan(recognizedText: receipt.text)
-                print("TEXT: " + receipt.text)
+                if receipt.text.count < 2 {
+                    return false
+                } else {
+                    Receipt.saveScan(recognizedText: receipt.text)
+                    return true
+                }
             }
             scanSelection = .none
         }
+        return false
     }
 }
 
@@ -87,59 +174,6 @@ struct ImagePicker: UIViewControllerRepresentable {
         
         func imagePickerViewController(_ controller: UIImagePickerController, didFailWithError error: Error) {
             galleryView.didFinishScanning(.failure(error))
-        }
-    }
-}
-
-// CAMERA SCANNER
-struct DocumentScannerView: View {
-    @Binding var scanSelection: ScanSelection
-    @Binding var isRecognizing: Bool
-    @ObservedObject var recognizedContent: RecognizedContent  = RecognizedContent()
-    
-    var body: some View {
-        if !UIDevice.current.isSimulator { // if device is physical (supports camera)
-            DocumentScanner { result in
-                switch result {
-                    case .success(let scannedImages):
-                        isRecognizing = true
-                        print(recognizedContent.items.count)
-                        TextRecognition(scannedImages: scannedImages, recognizedContent: recognizedContent) {
-                            saveReceipt()
-                            isRecognizing = false // Text recognition is finished, hide the progress indicator.
-                            print(recognizedContent.items[0])
-                        }.recognizeText()
-                    case .failure(let error):
-                        print(error.localizedDescription)
-                }
-                scanSelection = .none
-            } didCancelScanning: {
-                // Dismiss the scanner
-                scanSelection = .none
-            }
-        } else { // else if its in the simulator (no camera)
-            Text("Not supported in the simulator!\n\nPlease use a physical device.")
-                .font(.system(.title, design: .rounded))
-                .padding()
-            Button(action: {
-                scanSelection = .none
-            }){
-                Text("BACK")
-                    .padding().padding(.horizontal)
-                    .background(RoundedRectangle(cornerRadius: 25).fill(Color("object")))
-                    .padding()
-            }.buttonStyle(ShrinkingButton())
-            Spacer()
-        }
-    }
-    
-    func saveReceipt(){
-        if recognizedContent.items.count > 0 {
-            for receipt in recognizedContent.items {
-                Receipt.saveScan(recognizedText: receipt.text)
-                print("TEXT: " + receipt.text)
-            }
-            scanSelection = .none
         }
     }
 }
